Fixed #5 and #9 I didn't want to make such a large commit but here it is.
PORT_SERVER is not an attribute from Config anymore (this class wasn't
using it) but from RPC, which really uses it.

Also, Config is accessed from Client as an object.

Client has a support class ClientFile to store info about the files
passed as aguments (filepath, filename, etc).
Due to this change a lot of variable names have been modified.

-p and -g handles these file infos differently. When sending the file
the sytem checks if it exists, if it's a regular file and if it's in the
mapped directory. This is to avoid messy server databases, files being
created outise FlexA dir and also makes the receiving process easier.
NOTE: Later -g will check if the parameter is outside mapped dir too.

Finally, -p now skips these pesky arguments.