Added the field checksum to File entity
Checksum of the latest version is stored with its metadata, then, uppon
a write request (-p) it compares the checksum of the file to be sent
with the checksum stored.
If they are the same it means that the files have the same content, so
the only thing updated is the file metadata, but not the parts metadata
nor the data itself.